Refactor : Text 컴포넌트 관련 스타일 및 타입 파일 분리 (#19)
